rename: infer_batch -> schedule_batch
